requirements.txt
python-discord
python-dotenv
openai
